+++ src/main/java/com/spinner/www/myprofile/service/MyProfileService.java
@@ -34,16 +34,6 @@ public interface MyProfileService {
     ResponseEntity<CommonResponse> getSliceOfBookmarkedBoard(Long idx, int size);
 
     /**
-     * 인기글 게시글 목록 조회
-     * @param boardType String 조회할 게시판의 타입
-     * @param idx Long 조회 시작 idx
-     * @param size int 조회할 목록 갯수
-     * @return ResponseEntity<CommonResponse> 게시글 목록
-     */
-    ResponseEntity<CommonResponse> getSliceOfHotBoard(String boardType, Long idx, int size);
-
-
-    /**
      * idx로 회원 프로필 조회
      * @return MemberProfile
      */
