+++ src/main/java/com/spinner/www/myprofile/service/MyProfileServiceImpl.java
@@ -72,18 +72,6 @@ public class MyProfileServiceImpl implements MyProfileService {
     }
 
     /**
-     * 인기글 게시글 목록 조회
-     * @param boardType String 조회할 게시판의 codeIdx
-     * @param idx Long 조회 시작 idx
-     * @param size int 조회할 목록 갯수
-     * @return ResponseEntity<CommonResponse> 게시글 목록
-     */
-    @Override
-    public ResponseEntity<CommonResponse> getSliceOfHotBoard(String boardType, Long idx, int size) {
-        return boardService.getSliceOfHotBoard(boardType,idx,size);
-    }
-
-    /**
      * 회원 프로필 조회
      * @return Member
      */
